@@ -6,21 +6,21 @@
     "start": "ts-node app.ts"
   },
   "dependencies": {
-    "@types/node": "^20.10.1",
-    "config": "^3.3.9",
+    "@types/node": "^20.12.8",
+    "config": "^3.3.11",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
     "express": "^4.19.2",
     "express-fileupload": "^1.5.0",
-    "gif-frames": "^1.0.1",
+    "gif-extract-frames": "^1.0.2",
     "helmet": "^7.1.0",
     "lodash": "^4.17.21",
     "mongoose": "^8.3.3",
     "morgan": "^1.10.0",
     "pngjs": "^7.0.0",
     "sharp": "^0.33.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
@@ -30,9 +30,8 @@
     "@types/express": "^4.17.21",
     "@types/express-fileupload": "^1.5.0",
     "@types/lodash": "^4.17.0",
-    "@types/mongoose": "^5.11.97",
     "@types/morgan": "^1.9.9",
-    "@types/pngjs": "^6.0.4",
+    "@types/pngjs": "^6.0.5",
     "@types/uuid": "^9.0.8"
   }
 }
